feat(runtime): extract /packages dispatcher domain body — ADR-0076 D11 step ③ PR-5 (#2462)
The largest extraction so far: handlePackages (~470 lines) plus its two
EXCLUSIVE helpers assemblePackageManifest and applyPublishedSeeds
(~190 more) move to domains/packages.ts. resolveActiveOrganizationId
stays on the dispatcher (metadata-domain call sites share it) and is
exposed through deps instead.

DomainHandlerDeps grows: errorFromThrown (spec-validation issues carry
through as field-anchored 422s), resolveActiveOrganizationId,
announceKernelEvent (the metadata:reloaded publish announcement — the
domain keeps the changed-list check, the deps side owns the
trigger-existence check), and optional logger (seed-loader falls back
to console as before). The dynamic './seed-loader.js' import moved to
'../seed-loader.js' (the PR#2415 lesson: dynamic imports are the
extraction landmine — this time the import moved WITH the caller).

Step-② (#3142) single-pipeline behavior preserved; the '/packages'
bare-startsWith match is reproduced as-is.

Verified: seam suite 34 tests (4 new incl. the 409 duplicate-install
guard + ?overwrite=true), runtime 639 green, http-conformance 41 green,
dependent closure builds with DTS (--force).
